+++ homeassistant/components/decora_wifi/__init__.py
@@ -3,7 +3,11 @@
 import logging
 
 from decora_wifi import DecoraWiFiSession
+from decora_wifi.models.iot_switch import IotSwitch
+from decora_wifi.models.permission import Permission
 from decora_wifi.models.person import Person
+from decora_wifi.models.residence import Residence
+from decora_wifi.models.residential_account import ResidentialAccount
 import voluptuous as vol
 
 from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
@@ -78,3 +82,40 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     unload_ok = await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
     hass.data.pop(DOMAIN)
     return unload_ok
+
+
+class DecoraWifiAsyncClient:
+    """A thin wrapper to make async calls more readable."""
+
+    def __init__(self, session: DecoraWiFiSession, hass: HomeAssistant) -> None:
+        """DecoraWifiAsyncClient."""
+        self.session = session
+        self.hass = hass
+
+    async def get_permissions(self):
+        """Get all permissions for the provided session."""
+        assert self.session.user
+        return await self.hass.async_add_executor_job(
+            self.session.user.get_residential_permissions
+        )
+
+    async def get_residences(self, permissions: list[Permission]):
+        """Get all residences for the provided permissions."""
+        residences: list[Residence] = []
+        for perm in permissions:
+            if perm.residentialAccountId is not None:
+                account = ResidentialAccount(self.session, perm.residentialAccountId)
+                residences.extend(
+                    await self.hass.async_add_executor_job(account.get_residences)
+                )
+            elif perm.residenceId is not None:
+                residences.append(Residence(self.session, perm.residenceId))
+        return residences
+
+    async def get_iot_switches(self, residences: list[Residence]) -> list[IotSwitch]:
+        """Get all the iot switches for the provided residences."""
+        return [
+            sw
+            for res in residences
+            for sw in (await self.hass.async_add_executor_job(res.get_iot_switches))
+        ]
