+++ homeassistant/components/decora_wifi/strings.json
@@ -5,8 +5,8 @@
         "title": "Setup Decora Wifi",
         "description": "Your my.leviton.com credentials",
         "data": {
-          "username": "Username",
-          "password": "Password"
+          "username": "[%key:common::config_flow::data::username%]",
+          "password": "[%key:common::config_flow::data::password%]"
         }
       }
     },
